push fix erro and no finish notif
