+++ package.json
@@ -12,6 +12,7 @@
     "axios": "^0.19.2",
     "clsx": "^1.1.0",
     "dotenv": "^8.2.0",
+    "install": "^0.13.0",
     "jwt-decode": "^2.2.0",
     "node-sass": "^4.13.1",
     "react": "^16.12.0",
@@ -19,6 +20,7 @@
     "react-i18next": "^11.3.3",
     "react-image-gallery": "^1.0.7",
     "react-images": "^1.1.7",
+    "react-notifications-component": "^2.4.0",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.0",
